+++ packages/ibm-products/src/components/Datagrid/Datagrid.stories.js
@@ -32,7 +32,7 @@ import styles from './_storybook-styles.scss';
 import { DatagridActions } from './utils/DatagridActions';
 import { DatagridPagination } from './utils/DatagridPagination';
 import { Wrapper } from './utils/Wrapper';
-import { DocsPage } from './Datagrid.docs-page';
+import DocsPage from './Datagrid.docs-page';
 
 export default {
   title: getStoryTitle(Datagrid.displayName),
@@ -487,7 +487,8 @@ export const BatchActions = () => {
       DatagridBatchActions,
       rowActions: getRowActions(),
       onRowSelect: (row, event) => console.log('onRowClick: ', row, event),
-      onAllRowSelect: (rows, event) => console.log('onAllRowsClick called', rows, event),
+      onAllRowSelect: (rows, event) =>
+        console.log('onAllRowsClick called', rows, event),
       batchActionMenuButtonLabel: 'More',
     },
     useSelectRows,
@@ -495,7 +496,12 @@ export const BatchActions = () => {
     useStickyColumn
   );
 
-  return <Datagrid datagridState={{ ...datagridState }} />;
+  return (
+    <Datagrid
+      datagridState={{ ...datagridState }}
+      ariaToolbarLabel="batch actions toolbar"
+    />
+  );
 };
 
 export const DisableSelectRow = () => {