+++ packages/ibm-products/src/components/Datagrid/Datagrid/DatagridRow.js
@@ -1,16 +1,17 @@
 /**
- * Copyright IBM Corp. 2020, 2023
+ * Copyright IBM Corp. 2020, 2024
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
 
-import React from 'react';
+import React, { isValidElement } from 'react';
 import { TableRow, TableCell, SkeletonText } from '@carbon/react';
 import { px } from '@carbon/layout';
 import { selectionColumnId } from '../common-column-ids';
 import cx from 'classnames';
 import { pkg, carbon } from '../../../settings';
+import { DatagridSlug } from './addons/Slug/DatagridSlug';
 
 const blockClass = `${pkg.prefix}--datagrid`;
 
@@ -26,6 +27,7 @@ const rowHeights = {
 const DatagridRow = (datagridState) => {
   const {
     row,
+    rows,
     rowSize,
     withNestedRows,
     prepareRow,
@@ -34,6 +36,7 @@ const DatagridRow = (datagridState) => {
     withExpandedRows,
     withMouseHover,
     setMouseOverRowIndex,
+    headers,
   } = datagridState;
 
   const getVisibleNestedRowCount = ({ isExpanded, subRows }) => {
@@ -114,12 +117,6 @@ const DatagridRow = (datagridState) => {
     }
   };
 
-  const rowClassNames = cx(`${blockClass}__carbon-row`, {
-    [`${blockClass}__carbon-row-expanded`]: row.isExpanded,
-    [`${blockClass}__carbon-row-expandable`]: row.canExpand,
-    [`${carbon.prefix}--data-table--selected`]: row.isSelected,
-  });
-
   const setAdditionalRowProps = () => {
     if (withNestedRows || withExpandedRows) {
       return {
@@ -130,13 +127,21 @@ const DatagridRow = (datagridState) => {
   };
 
   // eslint-disable-next-line no-unused-vars
-  const { role, ...rowProps } = row.getRowProps();
+  const { role, className, ...rowProps } = row.getRowProps();
+  const foundAIRow = rows.some((r) => isValidElement(r?.original?.slug));
+
+  const rowClassNames = cx(`${blockClass}__carbon-row`, {
+    [`${blockClass}__carbon-row-expanded`]: row.isExpanded,
+    [`${blockClass}__carbon-row-expandable`]: row.canExpand,
+    [`${carbon.prefix}--data-table--selected`]: row.isSelected,
+    [`${blockClass}__slug--row`]: isValidElement(row?.original?.slug),
+  });
 
   return (
     <React.Fragment key={key}>
       <TableRow
-        className={rowClassNames}
         {...rowProps}
+        className={cx(rowClassNames, className)}
         key={row.id}
         onMouseEnter={hoverHandler}
         onMouseLeave={handleMouseLeave}
@@ -145,6 +150,19 @@ const DatagridRow = (datagridState) => {
         onKeyUp={handleOnKeyUp}
         {...setAdditionalRowProps()}
       >
+        {foundAIRow ? (
+          row?.original?.slug ? (
+            <td
+              className={cx(`${blockClass}__table-row-ai-enabled`, {
+                [`${blockClass}__slug--expanded`]: row.isExpanded,
+              })}
+            >
+              <DatagridSlug slug={row?.original?.slug} />
+            </td>
+          ) : (
+            <td className={`${blockClass}__table-row-ai-spacer`} />
+          )
+        ) : null}
         {row.cells.map((cell, index) => {
           const cellProps = cell.getCellProps();
           // eslint-disable-next-line no-unused-vars
@@ -160,11 +178,18 @@ const DatagridRow = (datagridState) => {
             return cell.render('Cell', { key: cell.column.id });
           }
           const title = content?.props?.children[0]?.props?.value;
+          const associatedHeader = headers.filter(
+            (h) => h.id === cell.column.id
+          );
           return (
             <TableCell
               className={cx(`${blockClass}__cell`, {
                 [`${blockClass}__expandable-row-cell`]:
                   row.canExpand && index === 0,
+                [`${blockClass}__slug--cell`]:
+                  associatedHeader &&
+                  associatedHeader.length &&
+                  isValidElement(associatedHeader[0]?.slug),
               })}
               {...restProps}
               key={cell.column.id}